Add metric for quorum queue availability
The per-object metric returns 1 for up, and 0 for down.
The aggregated metric returns the sum of quorum queues having a leader
on this node for which the leader reports that a majority is available
in the Raft cluster.

This change enables us to define alerts if quorum queues are down and to
define SLOs on quorum queue availability.